web系统/cgi-bin/requirements.txt
pymysql==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pip==23.2.1
setuptools==60.2.0
urllib3==2.0.4
wheel==0.37.1
colorlog==6.7.0
lxml==4.9.3
Requests==2.31.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.4
dnspython==2.2.1
docopt==0.6.2
exrex==0.10.5
fire==0.3.1
future==0.18.2
idna==3.3
loguru==0.6.0
mysql-connector-python==8.1.0
pipreqs==0.4.13
protobuf==4.21.12
Pygments==2.16.1
PySocks==1.7.1
python-fire==0.1.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2
SQLAlchemy==1.3.22
tenacity==8.0.1
termcolor==1.1.0
tqdm==4.64.0
treelib==1.6.1
urllib3==1.26.9
win32-setctime==1.1.0
yarg==0.1.9